requirements.txt
fastapi
uvicorn
prometheus-client
faiss-cpu
sentence-transformers
numpy
